+++ src/gitmit/utils/args.py
@@ -65,6 +65,12 @@ def __commit_parser(subparsers: argparse._SubParsersAction):
     )
 
     parser.add_argument(
+        "--debug",
+        help="Display debug information, such as files that will be committed and current prompt.",
+        action="store_true",
+    )
+
+    parser.add_argument(
         "-m",
         "--mode",
         help="Set the mode of the commit.",
@@ -204,7 +210,9 @@ def parse_args(version: str):
 
     subparsers = parser.add_subparsers(dest="command", help="Command to be executed.")
 
-    subparsers.add_parser("config", help="Display configuration and manage config file.")
+    subparsers.add_parser(
+        "config", help="Display configuration and manage config file."
+    )
     __commit_parser(subparsers)
     __init_parser(subparsers)
     __merge_parser(subparsers)