+++ CHANGELOG.md
@@ -29,4 +29,6 @@
 
 - [Add] Add `.gitmitignore` file to support to exclude files from commit prompt;
 - [Add] Add `--no-feat` option to ignore `feat` commit type at commit prompt;
-- [Changes] Commit prompt improvements.
+- [Add] Add `--debug` option on commit command to display files that will be committed and current prompt;
+- [Changes] Config command improvements, it displays all configuration (and usage) in a panel;
+- [Changes] Commit prompt improvements;